+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- First draft of CLI to check module compat with certain Magento version
+
 ### Fixed 
 - Make sure that non-namespaced classes are picked up as well (#30)
 - Reimplement flag "hide-needless" again